pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "saegenrec"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "YRW" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "datasets>=2.14",
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "mkdocs",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "pyyaml>=6.0",
    "requests>=2.28",
    "ruff",
    "scikit-learn",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "tqdm",
    "transformers>=4.30",
    "typer",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["saegenrec"]
include = ["pyproject.toml", "saegenrec/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["saegenrec"]
force-sort-within-sections = true

